requirements.txt
pyyaml
CodernityDB==0.4.2
jinja==1.2
configure==0.5
python3-ldap
sphinx